Moving evaluators to mathics.eval module.
   * adding the module mathics.eval

   * mathics.core.evaluators-> mathics.eval.nevaluator

   * formatter.format_element and related routines -> mathics.eval.makeboxes
