+++ mathics/builtin/binary/io.py
@@ -14,7 +14,6 @@ from mathics.builtin.base import Builtin
 from mathics.core.atoms import Complex, Integer, MachineReal, Real, String
 from mathics.core.convert.expression import to_expression, to_mathics_list
 from mathics.core.convert.mpmath import from_mpmath
-from mathics.core.evaluators import eval_N
 from mathics.core.expression import Expression
 
 from mathics.core.systemsymbols import (
@@ -27,6 +26,8 @@ from mathics.core.read import SymbolEndOfFile
 from mathics.core.streams import stream_manager
 from mathics.core.symbols import Symbol
 from mathics.core.systemsymbols import SymbolComplex
+
+from mathics.eval.nevaluator import eval_N
 
 SymbolBinaryWrite = Symbol("BinaryWrite")
