+++ mathics/builtin/drawing/plot.py
@@ -33,7 +33,6 @@ from mathics.core.atoms import (
 from mathics.core.attributes import A_HOLD_ALL, A_PROTECTED
 from mathics.core.convert.expression import to_expression, to_mathics_list
 from mathics.core.convert.python import from_python
-from mathics.core.evaluators import eval_N
 from mathics.core.expression import Expression
 from mathics.core.list import ListExpression
 from mathics.core.symbols import Symbol, SymbolList, SymbolN, SymbolPower, SymbolTrue
@@ -56,6 +55,8 @@ from mathics.core.systemsymbols import (
     SymbolSlot,
     SymbolStyle,
 )
+
+from mathics.eval.nevaluator import eval_N
 
 RealPoint6 = Real(0.6)
 RealPoint2 = Real(0.2)
