+++ mathics/core/element.py
@@ -246,8 +246,8 @@ class BaseElement(KeyComparable):
         return None
 
     def format(self, evaluation, form, **kwargs) -> "BoxElementMixin":
-        from mathics.core.formatter import format_element
         from mathics.core.symbols import Symbol
+        from mathics.eval.makeboxes import format_element
 
         if isinstance(form, str):
             form = Symbol(form)
